feat: Add new drone training courses and update course links
- Introduced new pages for Radio Telephony, Remote Pilot License, and Thermography courses.
- Updated courseList in constants to include slugs for new courses.
- Modified course card links to use slugs instead of titles for better URL structure.
- Enhanced the training links to reflect the new slug-based URLs.
- Added learning outcomes for Remote Pilot License and Radio Telephony courses.
