+++ packages/constants/src/courses.ts
@@ -416,24 +416,28 @@ export default courses;
 
 export const courseList = [
   {
+    slug: 'remote-pilot-license',
     image: images.training.rpl,
     linkDesc: 'Earn your RPL to operate drones professionally.',
     title: 'Remote Pilot License (RPL)',
     desc: 'The RPL course is designed for individuals who are interested in starting a career as a drone pilot or for unlicensed drone pilots who want to fly legally. This course is also suitable for individuals who are looking to add drone operation skills to their CVs. By obtaining an RPL, you will be able to legally operate drones in Kenya and potentially pursue a career in various industries, including cinematography, agriculture, and construction, survey and mapping among others.',
   },
   {
+    slug: 'radio-telephony',
     image: images.training.radTel,
     linkDesc: 'Learn radio communication for controlled airspace operations.',
     title: 'Radio Telephony',
     desc: 'The Radio Telephony License course, offered in partnership with the Nairobi Flight Academy, is designed for Remote Pilot License (RPL) holders seeking to operate drones in controlled airspace and conduct Beyond Visual Line of Sight (BVLOS) operations. This comprehensive training equips participants with the necessary skills in radio telephony procedures and communication protocols. Upon passing the exam administered by the Kenya Civil Aviation Authority (KCAA), successful participants are granted a Radio Telephony Operator License valid for two years, renewable upon expiration.',
   },
   {
+    slug: 'instructor-rating',
     image: images.training.insRat,
     linkDesc: 'Get certified to train and mentor future drone pilots.',
     title: 'Instructor Rating',
     desc: 'The Instructor Rating course is designed for individuals who already hold an RPL and are interested in becoming certified drone instructors. This course is ideal for RPL holders who have a passion for teaching and want to share their knowledge and experience with others. By completing this course, students will learn how to effectively instruct RPL students, create course materials, and design training programs. Upon completion, students will receive an instructor rating from the Kenya Civil Aviation Authority (KCAA), which will enable them to train and certify future RPL holders.',
   },
   {
+    slug: 'mapping-and-survey',
     image: images.training.mapSur,
     linkDesc: 'Hands-on training in mapping, surveying, and data analysis.',
     title: 'Drone Mapping and Survey',
@@ -441,6 +445,7 @@ export const courseList = [
     desc: 'The Drone Mapping and Survey course is a 5-day course designed for RPL holders who want to learn how to use drones for mapping and surveying purposes. This course is suitable for engineers, surveyors, environmental scientists, among others. It is also ideal for individuals interested in starting a drone mapping and surveying business. By completing this course, students will acquire practical skills in drone operation, surveying techniques, and data analysis. They will also learn how to process and interpret data obtained from drone surveys to generate accurate maps and 3D models. ',
   },
   {
+    slug: 'thermography',
     image: images.training.thermography,
     linkDesc: 'ITC Level I training for inspection and diagnostics.',
     title: 'Thermography (ITC - Level I)',
@@ -448,15 +453,30 @@ export const courseList = [
     desc: 'The Level I Thermography Certification course is designed for professionals who want to develop practical skills in using thermography for inspections and diagnostics. This course is suitable for individuals working in industries such as electrical, mechanical, and building inspections, as well as professionals involved in research and development.',
   },
   {
+    slug: 'agricultural-spraying',
     image: images.training.agriSpray,
     linkDesc: 'Comprehensive training in agricultural drone spraying.',
     title: 'Agricultural Spraying',
     desc: `Drone Space Kenya offers a comprehensive 8-day Agricultural Spraying Course combining 3 days of theory and 5 days of practical training. The course equips participants with the skills to operate agricultural drones safely and effectively for precision spraying.`,
   },
   {
+    slug: 'bvlos',
     image: images.training.bvlos,
     linkDesc: 'Train to operate drones beyond visual line of sight (BVLOS).',
     title: 'Beyond Visual Line of Sight',
     desc: 'Drone Space Kenya’s Beyond Visual Line of Sight (BVLOS) Training Program is designed for certified Remote Pilots seeking to safely and effectively operate drones beyond visual range.',
   },
 ];
+
+export const outcomes = {
+  rpl: [
+    'Develop an understanding of drone regulations, safety procedures and drone technology.',
+    'Acquire skills in drone operation, including pre-flight checks, in-flight maneuvers, and emergency procedures.',
+    'Learn how to plan and execute drone missions.',
+  ],
+  radio: [
+    'Develop proficiency in radio telephony procedures for drone operations in controlled airspace and BVLOS missions.',
+    'Acquire comprehensive knowledge of radio phraseology, air traffic control communication, emergency procedures, and airspace regulations',
+    'Successfully pass the KCAA exam and obtain a two-year Radio Telephony Operator License, enabling compliant operations in controlled airspace.',
+  ],
+};
